API

Modules

LazyReports.LazyReports

Types and constants

LazyReports.LazyReport

Functions and macros

LazyReports.lazyreport

LazyReports.lazyreport!

LazyReports.lazyreport_for_show!

LazyReports.write_lazyreport

Documentation

LazyReports.LazyReports — Module

LazyReports

Lazy reports in Julia.

LazyReports.LazyReport — Type

struct LazyReport

Represents a lazy report.

Don't instantiate directly, use lazyreport()

LazyReports.lazyreport — Function

lazyreport()
lazyreport(contents...)
lazyreport(contents...)

Generate a lazy report, e.g. a data processing report.

Use lazyreport!(rpt, contents...) to add more content to a report.

Example:

using LazyReports, StructArrays, IntervalSets, Plots

tbl = StructArray(
    col1 = rand(5), col2 = ClosedInterval.(rand(5), rand(5).+1),
    col3 = [rand(3) for i in 1:5], col4 = rand(Bool, 5),
    col5 = [:a, :b, :c, :d, :e], col6 = ["a", "b", "c", "d", "e"],
    col7 = [:(a[1]), :(a[2]), :(a[3]), :(a[4]), :(a[5])]
)

rpt = lazyreport(
    "# New report",
    "Table 1:", tbl
)
lazyreport!(rpt, "Figure 1:", stephist(randn(10^3)))
lazyreport!(rpt, "Figure 2:", histogram2d(randn(10^4), randn(10^4), format = :png))

show(stdout, MIME"text/plain"(), rpt)
show(stdout, MIME"text/html"(), rpt)
show(stdout, MIME"text/markdown"(), rpt)

write_lazyreport("report.txt", rpt)
write_lazyreport("report.html", rpt)
write_lazyreport("report.md", rpt)

See LazyReports.lazyreport_for_show! for how to specialize the behavior of show for specific report content types.

LazyReports.lazyreport! — Function

lazyreport!(rpt::LazyReport, contents...)

Add more content to report rpt. See lazyreport for an example.

LazyReports.lazyreport_for_show! — Function

LazyReports.lazyreport_for_show!(rpt::LazyReport, mime::MIME, content)

Add the contents of content to rpt in a way that is optimized for being displayed (e.g. via show) with the given mime type.

show(output, mime, rpt) first transforms rpt by converting all contents of rpt using lazyreport_for_show!(rpt::LazyReport, mime, content).

Defaults to lazyreport!(rpt, content), except for tables (Tables.istable(content) == true), which are converted to Markdown tables by default for uniform appearance.

lazyreport_for_show! is not inteded to be called by users, but to be specialized for specific types of content content. Content types not already supported will primarily require specialization of

lazyreport_for_show!(rpt::LazyReport, ::MIME"text/markdown", content::SomeType)

In some cases it may be desireable to specialize lazyreport_for_show! for MIME types like MIME"text/html" and MIME"text/plain" as well.

LazyReports.write_lazyreport — Function

write_lazyreport(filename::AbstractString, rpt::LazyReport)
write_lazyreport(filename::AbstractString, mime::MIME, rpt::LazyReport)

Write lazyreport rpt to file filename.